Implements CBOR source generator
Adds a source generator for CBOR serialization and deserialization.
This allows for AOT compilation and avoids runtime reflection.

Introduces attributes to mark types and properties for serialization,
as well as options to configure the generated code.

Includes a demo project to showcase the usage of the generator.

Targets netstandard2.0 for broader compatibility.